fix all 30-ish compile time errors
