+++ Barkane/Assets/Scripts/Camera/FoldIndicator.cs
@@ -8,46 +8,46 @@ using UnityEngine;
 public class FoldIndicator : MonoBehaviour
 {
     [SerializeField] private GameObject ghostSquarePrefab;
-    public Vector2 foldCenter; //fold center in screen space
-    private List<Transform> transforms = new List<Transform>();
-    new private Camera camera;
+    // public Vector2 foldCenter; //fold center in screen space
+    // private List<Transform> transforms = new List<Transform>();
+    // new private Camera camera;
     private List<Vector3Int> locs = new List<Vector3Int>();
     
     public Vector3 Center = Vector3.zero;
 
-    private void Update() 
-    {
-        if(camera == null) return;
-        foldCenter = camera.WorldToScreenPoint(CoordUtils.CalculateCenterTransform(transforms));
-    }
+    // private void Update() 
+    // {
+    //     if(camera == null) return;
+    //     foldCenter = camera.WorldToScreenPoint(CoordUtils.CalculateCenterTransform(transforms));
+    // }
     
-    public void BuildIndicator(FoldData fd, Camera c)
-    {
-        Transform center = fd.foldObjects.foldJoints[0].transform;
-        transform.position = center.position;
-        locs = new List<Vector3Int>();
-        foreach(GameObject go in fd.foldObjects.foldSquares)
-        {
-            Vector3Int pos = Vector3Int.RoundToInt(go.transform.position);
-            if(!locs.Contains(pos)){
-            locs.Add(pos);
-            GameObject newSquare = Instantiate(ghostSquarePrefab, go.transform.position, go.transform.rotation);
-            var ghostSquareRenderer = newSquare.GetComponent<MeshRenderer>();
-            ghostSquareRenderer.sharedMaterial = VFXManager.Theme.GhostMat;
-            newSquare.transform.parent = gameObject.transform;
-            }
-        }
-        transform.RotateAround(center.position, center.rotation * Vector3.right, fd.degrees);
-        int children = transform.childCount;
-        transforms.Clear();
-        for (int i = 0; i < children; i++)
-            transforms.Add(transform.GetChild(i));
+    // public void BuildIndicator(FoldData fd, Camera c)
+    // {
+    //     Transform center = fd.foldObjects.foldJoints[0].transform;
+    //     transform.position = center.position;
+    //     locs = new List<Vector3Int>();
+    //     foreach(GameObject go in fd.foldObjects.foldSquares)
+    //     {
+    //         Vector3Int pos = Vector3Int.RoundToInt(go.transform.position);
+    //         if(!locs.Contains(pos)){
+    //         locs.Add(pos);
+    //         GameObject newSquare = Instantiate(ghostSquarePrefab, go.transform.position, go.transform.rotation);
+    //         var ghostSquareRenderer = newSquare.GetComponent<MeshRenderer>();
+    //         ghostSquareRenderer.sharedMaterial = VFXManager.Theme.GhostMat;
+    //         newSquare.transform.parent = gameObject.transform;
+    //         }
+    //     }
+    //     transform.RotateAround(center.position, center.rotation * Vector3.right, fd.degrees);
+    //     int children = transform.childCount;
+    //     transforms.Clear();
+    //     for (int i = 0; i < children; i++)
+    //         transforms.Add(transform.GetChild(i));
         
-        camera = c;
-        foldCenter = camera.WorldToScreenPoint(CoordUtils.CalculateCenterTransform(transforms));
-    }
+    //     camera = c;
+    //     foldCenter = camera.WorldToScreenPoint(CoordUtils.CalculateCenterTransform(transforms));
+    // }
 
-    public void BuildIndicator2(List<FoldableObject> fd, Camera c)
+    public void BuildIndicator(List<FoldableObject> fd, Camera c)
     {
         foreach(SquareStack s in fd)
         {
