+++ Barkane/Assets/Scripts/Expiremental/LevelTester2.cs
@@ -1,27 +1,27 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-using BarkaneEditor;
+// using System.Collections;
+// using System.Collections.Generic;
+// using UnityEngine;
+// using UnityEngine.SceneManagement;
+// using BarkaneEditor;
 
 
-public class LevelTester2 : MonoBehaviour
-{
-    public Level level;
-    public GameObject playerPrefab;
+// public class LevelTester2 : MonoBehaviour
+// {
+//     public Level level;
+//     public GameObject playerPrefab;
 
-    private void Start() {
-        Invoke("SpawnLevel", 0.02f);
-    }
+//     private void Start() {
+//         Invoke("SpawnLevel", 0.02f);
+//     }
 
-    public void SpawnLevel()
-    {
-        GameObject instantiatedLevel = Instantiate(level.levelObject, Vector3.zero, Quaternion.identity);
-        FoldablePaper paper = instantiatedLevel.GetComponent<FoldablePaper>();
-        Transform playerPos = paper.playerSpawn;
-        GameObject playerInstance= Instantiate(playerPrefab, playerPos.position, Quaternion.identity);
-        FollowTarget.Instance.SetTargetAndPosition(playerInstance.GetComponent<PlayerMovement>().cameraTrackingTransform);    
-        VFXManager.Instance.Refresh();
-        FindObjectOfType<TileSelector2>().ReloadReferences();
-    }
-}
+//     public void SpawnLevel()
+//     {
+//         GameObject instantiatedLevel = Instantiate(level.levelObject, Vector3.zero, Quaternion.identity);
+//         FoldablePaper paper = instantiatedLevel.GetComponent<FoldablePaper>();
+//         Transform playerPos = paper.playerSpawn;
+//         GameObject playerInstance= Instantiate(playerPrefab, playerPos.position, Quaternion.identity);
+//         FollowTarget.Instance.SetTargetAndPosition(playerInstance.GetComponent<PlayerMovement>().cameraTrackingTransform);    
+//         VFXManager.Instance.Refresh();
+//         FindObjectOfType<TileSelector2>().ReloadReferences();
+//     }
+// }
