+++ Barkane/Assets/Scripts/Foldable Paper/FoldObjects.cs
@@ -16,7 +16,7 @@ public class FoldObjects {
 
     public HashSet<SquareSide> SidesSet;
 
-    public OcclusionMap OcclusionMap = new OcclusionMap();
+    //public OcclusionMap OcclusionMap = new OcclusionMap();
 
     public FoldObjects() {
         foldSquares = new List<GameObject>();
@@ -88,119 +88,119 @@ public class FoldObjects {
         return CoordUtils.CalculateCenter(vectors);
     }
 
-    public void TransferToLocalOcclusionMap(Matrix4x4 encode, Matrix4x4 decode)
-    {
-        OcclusionMap.Clear();
-        PaperSquaresCache = new List<PaperSquare>();
-        foreach (GameObject ps in foldSquares)
-        {
-            PaperSquaresCache.Add(ps.GetComponent<PaperSquare>());
-        }
+    // public void TransferToLocalOcclusionMap(Matrix4x4 encode, Matrix4x4 decode)
+    // {
+    //     OcclusionMap.Clear();
+    //     PaperSquaresCache = new List<PaperSquare>();
+    //     foreach (GameObject ps in foldSquares)
+    //     {
+    //         PaperSquaresCache.Add(ps.GetComponent<PaperSquare>());
+    //     }
 
-        foreach (var ps in PaperSquaresCache)
-        {
-            ps.EjectFromGlobalQueue();
-            var center = Vector3Int.RoundToInt(encode.MultiplyPoint(ps.transform.position));
+    //     foreach (var ps in PaperSquaresCache)
+    //     {
+    //         ps.EjectFromGlobalQueue();
+    //         var center = Vector3Int.RoundToInt(encode.MultiplyPoint(ps.transform.position));
 
-            if (!OcclusionMap.ContainsKey(center))
-            {
-                // use constant encoder factory as local structure does not change during fold
-                var q = OcclusionQueue.MakeOcclusionQueue(center, ()=>encode);
-                if (q == null)
-                    throw new UnityException($"Local occlusion map entry could not be created { center }");
-                else
-                    q.Enqueue(ps);
-                OcclusionMap[center] = q;
-            }
-            else
-                OcclusionMap[center].Enqueue(ps);
+    //         if (!OcclusionMap.ContainsKey(center))
+    //         {
+    //             // use constant encoder factory as local structure does not change during fold
+    //             var q = OcclusionQueue.MakeOcclusionQueue(center, ()=>encode);
+    //             if (q == null)
+    //                 throw new UnityException($"Local occlusion map entry could not be created { center }");
+    //             else
+    //                 q.Enqueue(ps);
+    //             OcclusionMap[center] = q;
+    //         }
+    //         else
+    //             OcclusionMap[center].Enqueue(ps);
 
-            //Debug.DrawRay(
-            //    decode.MultiplyPoint(OcclusionMap[center].Offset),
-            //    decode.MultiplyVector(OcclusionMap[center].SpatialBasis.normal),
-            //    Color.cyan, 6);
-        }
+    //         //Debug.DrawRay(
+    //         //    decode.MultiplyPoint(OcclusionMap[center].Offset),
+    //         //    decode.MultiplyVector(OcclusionMap[center].SpatialBasis.normal),
+    //         //    Color.cyan, 6);
+    //     }
 
-        // Debug.Log(OcclusionMap);
-    }
+    //     // Debug.Log(OcclusionMap);
+    // }
 
 
-    public void MergeWithGlobalOcclusionMap(
-        OcclusionMap globalMap, Vector3 worldSpaceRotationRoot, Vector3 ccwAxis, Func<float, (Matrix4x4 encoder, Matrix4x4 decoder)> replay)
-    {
-        var (encoder1, decoder1) = replay(1);
+    // public void MergeWithGlobalOcclusionMap(
+    //     OcclusionMap globalMap, Vector3 worldSpaceRotationRoot, Vector3 ccwAxis, Func<float, (Matrix4x4 encoder, Matrix4x4 decoder)> replay)
+    // {
+    //     var (encoder1, decoder1) = replay(1);
 
-        foreach (var (local, oq) in OcclusionMap)
-        {
-            var worldSpacePos = Vector3Int.RoundToInt(decoder1.MultiplyPoint(local));
-            var worldSpaceUp = Vector3Int.RoundToInt(decoder1.MultiplyVector(oq.SpatialBasis.normal));
+    //     foreach (var (local, oq) in OcclusionMap)
+    //     {
+    //         var worldSpacePos = Vector3Int.RoundToInt(decoder1.MultiplyPoint(local));
+    //         var worldSpaceUp = Vector3Int.RoundToInt(decoder1.MultiplyVector(oq.SpatialBasis.normal));
 
-            var alignedToNegative = worldSpaceUp.x < 0 || worldSpaceUp.y < 0 || worldSpaceUp.z < 0;
+    //         var alignedToNegative = worldSpaceUp.x < 0 || worldSpaceUp.y < 0 || worldSpaceUp.z < 0;
 
-            // Matching position and direction
-            if (globalMap.ContainsKey(worldSpacePos))
-            {
-                bool approachFromPositive;
-                var radial = worldSpacePos - worldSpaceRotationRoot;
-                var matchingFactor = Vector3.Dot(radial, worldSpaceUp);
-                if (Mathf.Abs(matchingFactor) < 0.05f)
-                {
-                    // Case where the tile is on the folding plane
-                    approachFromPositive = Vector3.Dot(Vector3.Cross(radial, ccwAxis), alignedToNegative ? -worldSpaceUp : worldSpaceUp) > 0;
-                } else
-                {
-                    // Inambiguous case where the tile has some slanted radial direction against the rotation center
-                    // Always insert new things "inward" w.r.t. rotation radial dir.
-                    // ex.
-                    //     __
-                    //      |
-                    // |    |
-                    // |____| --> +x
-                    // ^ existing wall stays outward
-                    //  _____
-                    // ||   |
-                    // |____| --> +x
-                    //  ^ new wall comes inward
-                    // 
-                    // ... in this case the new wall is approaching from positive, since it's radial is facing -x
+    //         // Matching position and direction
+    //         if (globalMap.ContainsKey(worldSpacePos))
+    //         {
+    //             bool approachFromPositive;
+    //             var radial = worldSpacePos - worldSpaceRotationRoot;
+    //             var matchingFactor = Vector3.Dot(radial, worldSpaceUp);
+    //             if (Mathf.Abs(matchingFactor) < 0.05f)
+    //             {
+    //                 // Case where the tile is on the folding plane
+    //                 approachFromPositive = Vector3.Dot(Vector3.Cross(radial, ccwAxis), alignedToNegative ? -worldSpaceUp : worldSpaceUp) > 0;
+    //             } else
+    //             {
+    //                 // Inambiguous case where the tile has some slanted radial direction against the rotation center
+    //                 // Always insert new things "inward" w.r.t. rotation radial dir.
+    //                 // ex.
+    //                 //     __
+    //                 //      |
+    //                 // |    |
+    //                 // |____| --> +x
+    //                 // ^ existing wall stays outward
+    //                 //  _____
+    //                 // ||   |
+    //                 // |____| --> +x
+    //                 //  ^ new wall comes inward
+    //                 // 
+    //                 // ... in this case the new wall is approaching from positive, since it's radial is facing -x
 
-                    // Obviously, in real life you can have the top flap going outwards *or* inwards, but
-                    // here we want to simply it to only one case
+    //                 // Obviously, in real life you can have the top flap going outwards *or* inwards, but
+    //                 // here we want to simply it to only one case
 
-                    // Note the inward direction is always offsetted *against* the radial
-                    // Since axis is always on an offsetted joint, the distance to the new tile center is always nonzero
-                    // This means we won't get a zero vector and the dot product is usable
-                    approachFromPositive = matchingFactor < 0;
-                }
+    //                 // Note the inward direction is always offsetted *against* the radial
+    //                 // Since axis is always on an offsetted joint, the distance to the new tile center is always nonzero
+    //                 // This means we won't get a zero vector and the dot product is usable
+    //                 approachFromPositive = matchingFactor < 0;
+    //             }
 
-                // Debug.Log($"approach from: {(approachFromPositive ? "+" : "-")} { local } -> { worldSpacePos } should flip: {(alignedToNegative ? "true":"false")}");
+    //             // Debug.Log($"approach from: {(approachFromPositive ? "+" : "-")} { local } -> { worldSpacePos } should flip: {(alignedToNegative ? "true":"false")}");
 
-                if (approachFromPositive)
-                    // When approaching from positive, the new tiles (contents of the local occlusion map) covers the old tiles
-                    // This means they come *after* the original items in the merged queue
-                    globalMap[worldSpacePos].MergeApproachingFromPositive(
-                        alignedToNegative ? oq.MakeFlippedCopy() : oq);
-                else
-                    // Coming from a local occlusion map content goes *before* the global content
-                    globalMap[worldSpacePos].MergeApproachingFromNegative(
-                        alignedToNegative ? oq.MakeFlippedCopy() : oq);
-            }
-            else
-            {
-                // Insert local entry directly into global entry
-                // Flip to always using positive direction
-                globalMap[worldSpacePos] = alignedToNegative ? oq.MakeFlippedCopy() : oq;
-            }
+    //             if (approachFromPositive)
+    //                 // When approaching from positive, the new tiles (contents of the local occlusion map) covers the old tiles
+    //                 // This means they come *after* the original items in the merged queue
+    //                 globalMap[worldSpacePos].MergeApproachingFromPositive(
+    //                     alignedToNegative ? oq.MakeFlippedCopy() : oq);
+    //             else
+    //                 // Coming from a local occlusion map content goes *before* the global content
+    //                 globalMap[worldSpacePos].MergeApproachingFromNegative(
+    //                     alignedToNegative ? oq.MakeFlippedCopy() : oq);
+    //         }
+    //         else
+    //         {
+    //             // Insert local entry directly into global entry
+    //             // Flip to always using positive direction
+    //             globalMap[worldSpacePos] = alignedToNegative ? oq.MakeFlippedCopy() : oq;
+    //         }
 
-            globalMap[worldSpacePos].UpdateSpace(
-                alignedToNegative ? -worldSpaceUp : worldSpaceUp,
-                worldSpacePos,
-                OcclusionQueue.IdentityEncoder);
-            globalMap[worldSpacePos].UseAsGlobal();
+    //         globalMap[worldSpacePos].UpdateSpace(
+    //             alignedToNegative ? -worldSpaceUp : worldSpaceUp,
+    //             worldSpacePos,
+    //             OcclusionQueue.IdentityEncoder);
+    //         globalMap[worldSpacePos].UseAsGlobal();
 
-            // Debug.DrawRay(globalMap[worldSpacePos].center, globalMap[worldSpacePos].upwards * 2, Color.magenta, 3);
-        }
+    //         // Debug.DrawRay(globalMap[worldSpacePos].center, globalMap[worldSpacePos].upwards * 2, Color.magenta, 3);
+    //     }
 
-        // Debug.Log(globalMap);
-    }
+    //     // Debug.Log(globalMap);
+    // }
 }
