+++ Barkane/Assets/Scripts/Foldable Paper/FoldableObject.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e61014166e9ca234093e4abdc85b1bbb
+guid: 514c161bdd58cbc4abdc9a047b0391b2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
