+++ Barkane/Assets/Scripts/Misc/LevelTester.cs
@@ -1,7 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.SceneManagement;
 using BarkaneEditor;
