+++ Barkane/Assets/Scripts/Player/PlayerMovement.cs
@@ -111,8 +111,7 @@ public class PlayerMovement : MonoBehaviour
     {
         isMoving = true;
         
-        //TileSelector.Instance.DeselectJoint();
-        TileSelector2.Instance.DeselectJoint();
+        TileSelector.Instance.DeselectJoint();
 
         Vector3 curr = transform.position;
         Vector3 goal = targetPos.transform.position;
